+++ Invoice.API/Controllers/Entities/ProductController.cs
@@ -1,4 +1,5 @@
-﻿using Invoice.Domain.Entities;
+﻿using Invoice.Core.Dtos;
+using Invoice.Domain.Entities;
 using Invoice.Domain.Interfaces.Repositories;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
@@ -19,39 +20,78 @@ namespace Invoice.API.Controllers.Entities{
         public async Task<IActionResult> GetProducts()
         {
             var products = await _productRepository.GetAllAsync();
-            return Ok(products);
+            var productDTOs = MapToProductResponseDtoList(products);
+            return Ok(productDTOs);
         }
         
         [HttpGet("{id}")]
         public async Task<IActionResult> GetProduct(Guid id)
         {
             var product = await _productRepository.GetByIdAsync(id);
-            return Ok(product);
+            if (product == null)
+                return NotFound();
+            var productDto = MapToProductResponseDto(product);
+            return Ok(productDto);
         }
         
         [HttpPost]
-        public async Task<IActionResult> CreateProduct([FromBody] Product product)
+        public async Task<IActionResult> CreateProduct(ProductDto productDTO)
         {
+            var product = MapToProduct(productDTO);
             _productRepository.Add(product);
             _productRepository.Save();
-            return Ok(await _productRepository.GetByIdAsync(product.Id));
+            return CreatedAtAction(nameof(GetProduct), new { id = product.Id }, MapToProductResponseDto(product));
         }
         
-        [HttpPut]
-        public async Task<IActionResult> UpdateProduct([FromBody] Product product)
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UpdateProduct(Guid id, ProductDto productDTO)
         {
-            _productRepository.Update(product);
+            var existingProduct = await _productRepository.GetByIdAsync(id);
+            if(existingProduct == null)
+                return NotFound();
+            
+            existingProduct.Name = productDTO.Name;
+            existingProduct.Price = productDTO.Price;
+            
+            _productRepository.Update(existingProduct);
             _productRepository.Save();
-            return Ok(await _productRepository.GetByIdAsync(product.Id));
+            return Ok(MapToProductResponseDto(existingProduct));
         }
         
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteProduct(Guid id)
         {
             var product = await _productRepository.GetByIdAsync(id);
+            if (product == null)
+                return NotFound();
             _productRepository.Delete(product);
             _productRepository.Save();
-            return Ok();
+            return NoContent();
+        }
+        
+        private ProductResponseDto MapToProductResponseDto(Product product)
+        {
+            return new ProductResponseDto
+            {
+                Id = product.Id,
+                Name = product.Name,
+                Price = product.Price
+            };
+        }
+        
+        private Product MapToProduct(ProductDto productDto)
+        {
+            return new Product
+            {
+                Id = productDto.Id,
+                Name = productDto.Name,
+                Price = productDto.Price
+            };
+        }
+        
+        private List<ProductResponseDto> MapToProductResponseDtoList(IEnumerable<Product> products)
+        {
+            return products.Select(MapToProductResponseDto).ToList();
         }
     }
